requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sse-starlette>=2.1.3,<3.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
sqlalchemy>=2.0.36,<3.0.0
asyncpg>=0.30.0,<1.0.0
aiosqlite>=0.20.0,<1.0.0
langchain>=0.3.13,<0.4.0
langchain-openai>=0.2.14,<0.3.0
langgraph>=0.2.60,<0.3.0
markdown>=3.7,<4.0
pypdf>=5.1.0,<6.0.0
python-docx>=1.1.2,<2.0.0
httpx>=0.28.1,<1.0.0
pytest>=8.3.4,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
ruff>=0.8.4,<1.0.0
